This is a big changeset that turns upside down a lot of how the parts within the UPnP framework are dealt with.
There is now an embracement class called Coherence that
deals with all the parts needed to run an UPnP service.
And allows in the future a more plugin based structure.

There we decide whether we only want to be a ControlPoint,
only a MediaRenderer, only a MediaServer or any combination 
of these.

The SSDP and the MSEARCH subsystems have now their home there.

And only one webserver instance is created where all the other
subsystems can register their services as childs.

For example the ControlPoint appends there its Event Notification
service and its XMLRPC interface.
And of course it can append there its own web-interface.

Likewise for the MediaServer. It appends its XML files,
the Event Subscription service, the SOAP interface and its
media files.

Hopefully not to many things broke with this changeset,
actually and intendedly we should be on the same functionality
level like before.