A lot of fixes and enhancements regarding a working MediaServer.
Some of them concerning the different namespace, which refs to #28.

Fixed an embarrassingly stupid bug in the device registration, where we had the services url in the location header of the service notification, instead of the url for the device description.xml, this is one of those that fixes #26.

This only showed up due to the fact that the Noxon-2 looks only for service notifications.

Version 2 devices now send out notifications for devices v1 and services v1 too. Requisite for this are different device descriptions accordingly to the version, pointing to the matching services.