Check if a device backend for server service provides all needed action callbacks.
If the action is optional, just don't provide it.
If it is a required action, don't activate the service.

If that service is a required one, this will lead to an abort of the device
instantiation.

We need an overwrite for this to create in the TestSuite nonconformity devices/services.

This closes #25.