Fix the shutdown method of Coherence, the ''reactor.addSystemEventTrigger( 'before', 'shutdown', self.shutdown)'' expects self.shutdown to return a deferred and waits until that has finished.
Now we return a [http://twistedmatrix.com/projects/core/documentation/howto/defer.html#auto7 deferredList] in which we collect and wait for all cleanups that have to be done.

In addition the way the event unsubscription had to be changed. A new EventProtocol instance seems not to be created by the ClientCreater when the reactor is going down. Which makes some sense anyway. :-)

So we keep the protocol instance after our first event subscription request for reuse around and are now able to send out the unsubscribe request.

This fixes #75 - at least it is supposed to. If not pls reopen this ticket.