enable support for device icons
configuration isn't yet optimal -  in the section of the MediaServer or MediaRenderer backend in the config file we need something like this:

{{{
[[[icon]]]
        mimetype = image/png
        height = 48
        width = 48
        depth = 24
        url = file:///home/dev/elisa/elisa.png
}}}

or in the config dict:

{{{
'icon': { 'url':'http://coherence.beebits.net/image.png',
             'mimetype': 'image/png',
             'width':120,
             'height':120,
             'depth':24
          }
}}}

According to the [http://upnp.org/specs/arch/UPnP-DeviceArchitecture-v1.0-20060720.pdf UPnP Device Architecture p.28] all fields are required. Maybe later Coherence can autogenerate the values with a PIL call.

Multiple icons are supported to. The config above has to be wrapped into an '''icons''' section or dictionary, and the '''icon''' has to be replaced with an unique name.