Security fix: FSStore exports whole user-dir if passed no pass at all.
When an empty content is passed to update_config() in FSStore backend
(for example when all content is removed by the export widget in
av_widgets.py), the os.path.abspath("") expands to the user's home
folder thus exporting all his folders and files there. This is totally
unacceptable and is now fixed.

This solves one part of issue #338. Thanks to PhobosK for the fix.