Security fix: FSStore by default shares Music, Videos, Pictures folders.
The problem is that the user is not aware of this behaviour and these
folders could contain sensitive materials that the user does not want
to share.

Solution: It is much better to use as default share folder the
XDG_PUBLICSHARE_DIR. At least the user knows it is shared and he does
not keep sensitive materials there.

This solves the other part of issue #338. Thanks to PhobosK for the fix.